+++ src/main/java/net/momostudios/coldsweat/core/capabilities/ITemperatureCapability.java
@@ -4,6 +4,6 @@ import net.momostudios.coldsweat.core.util.PlayerTemp;
 
 public interface ITemperatureCapability
 {
-    double get(PlayerTemp.Types type);
-    void set(PlayerTemp.Types type, double value);
+    float get(PlayerTemp.Types type);
+    void set(PlayerTemp.Types type, float value);
 }
